Don't require "--network nil" parameter for hardhat examples
Currently we don't set the default network in hardhat configuration,
and because of this users have to do:

```
npx hardhat ignition deploy ignition/modules/<module>.ts --network nil
```

This is not convenient, because obviously they mean to always use our
network with our examples. So we can just skip `--network nil` if we
set up a default network.